requirements.txt
uvicorn==0.18.3
fastapi==0.75.0
cryptography==38.0.1
python-socketio==4.6.1
python-jose==3.3.0
python-multipart==0.0.5
shapely==1.8.4
eventlet==0.33.1
passlib[bcrypt]==1.7.4
psutil==5.9.2
jinja2==3.0.3
aiofiles==0.8.0
imutils==0.5.4
openvino-dev
numpy==1.19.4
torch==1.8.1
torchvision==0.9.1
matplotlib==3.6.0
opencv-python
loguru
scikit-image
tqdm
Pillow==9.0.0
thop
ninja
tabulate
tensorboard
python-dotenv
pyinstaller
auto-py-to-exe==2.22.0
paddlepaddle==2.5.1
pyclipper==1.2.1
ipywidgets
IPython
ultralytics